+++ src/HttpClient/HttpClient.php
@@ -106,7 +106,9 @@ class HttpClient implements HttpClientInterface
         if ($body === false) {
             $errorCode = curl_errno($curlHandle);
             $error = curl_error($curlHandle);
-            curl_close($curlHandle);
+            if (\PHP_MAJOR_VERSION < 8) {
+                curl_close($curlHandle);
+            }
 
             $message = 'cURL Error (' . $errorCode . ') ' . $error;
 
@@ -115,7 +117,9 @@ class HttpClient implements HttpClientInterface
 
         $statusCode = curl_getinfo($curlHandle, \CURLINFO_HTTP_CODE);
 
-        curl_close($curlHandle);
+        if (\PHP_MAJOR_VERSION < 8) {
+            curl_close($curlHandle);
+        }
 
         $error = $statusCode >= 400 ? $body : '';