+++ src/Tracing/Span.php
@@ -7,6 +7,7 @@ namespace Sentry\Tracing;
 use Sentry\EventId;
 use Sentry\SentrySdk;
 use Sentry\State\Scope;
+use Sentry\Unit;
 
 /**
  * This class stores all the information about a span.
@@ -21,6 +22,13 @@ use Sentry\State\Scope;
  */
 class Span
 {
+    /**
+     * Maximum number of flags allowed. We only track the first flags set.
+     *
+     * @internal
+     */
+    public const MAX_FLAGS = 10;
+
     /**
      * @var SpanId Span ID
      */
@@ -60,6 +68,11 @@ class Span
      * @var array<string, string> A List of tags associated to this span
      */
     protected $tags = [];
+
+    /**
+     * @var array<string, bool> A List of flags associated to this span
+     */
+    protected $flags = [];
 
     /**
      * @var array<string, mixed> An arbitrary mapping of additional metadata
@@ -328,6 +341,20 @@ class Span
     }
 
     /**
+     * Sets a feature flag associated to this span.
+     *
+     * @return $this
+     */
+    public function setFlag(string $key, bool $result)
+    {
+        if (\count($this->flags) < self::MAX_FLAGS) {
+            $this->flags[$key] = $result;
+        }
+
+        return $this;
+    }
+
+    /**
      * Gets the ID of the span.
      */
     public function getSpanId(): SpanId
@@ -368,7 +395,13 @@ class Span
     public function getData(?string $key = null, $default = null)
     {
         if ($key === null) {
-            return $this->data;
+            $data = $this->data;
+
+            foreach ($this->flags as $flagKey => $flagValue) {
+                $data["flag.evaluation.{$flagKey}"] = $flagValue;
+            }
+
+            return $data;
         }
 
         return $this->data[$key] ?? $default;